Nemesis Transfer Protocol (Win32, Linux and OS/2)
Nemesis Transfer Protocol (NTP) is a file transfer client / server set
used with BBS Programs in Windows, Linux, and OS/2.

The user would start up the user end (server) applications.  Once
connected to the bbs, they would select to download with NTP protocol.
The bbs would when pass the socket to the BBS client end which would
connect and transfer files directly to the end user.